Extract `TempPath` and `NonLockingFileReader` utility classes for reuse across tests, and refactor `RollingLogger` tests to use the shared implementations.
