+++ src/NServiceBus.Core.Tests/Helpers/NonLockingFileReader.cs
@@ -1,0 +1,13 @@
+namespace NServiceBus.Core.Tests.Helpers;
+
+using System.IO;
+
+static class NonLockingFileReader
+{
+    internal static string ReadAllTextWithoutLocking(string path)
+    {
+        using var fileStream = new FileStream(path, FileMode.Open, FileAccess.Read, FileShare.ReadWrite);
+        using var textReader = new StreamReader(fileStream);
+        return textReader.ReadToEnd();
+    }
+}
