+++ src/NServiceBus.Core.Tests/Logging/RollingLoggerTests.cs
@@ -3,7 +3,7 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
+using Helpers;
 using NUnit.Framework;
 
 [TestFixture]
@@ -12,7 +12,7 @@ public class RollingLoggerTests
     [Test]
     public void When_file_already_exists_that_file_is_written_to()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var dateTime = new DateTimeOffset(2010, 10, 1, 0, 0, 0, TimeSpan.Zero);
         var logger1 = new RollingLogger(tempPath.TempDirectory)
         {
@@ -23,7 +23,7 @@ public class RollingLoggerTests
         using (Assert.EnterMultipleScope())
         {
             Assert.That(files1, Has.Count.EqualTo(1));
-            Assert.That(NonLockingFileReader.ReadAllTextWithoutLocking(files1.First()), Is.EqualTo($"Foo{Environment.NewLine}"));
+            Assert.That(NonLockingFileReader.ReadAllTextWithoutLocking(files1[0]), Is.EqualTo($"Foo{Environment.NewLine}"));
         }
         var logger2 = new RollingLogger(tempPath.TempDirectory)
         {
@@ -34,14 +34,14 @@ public class RollingLoggerTests
         using (Assert.EnterMultipleScope())
         {
             Assert.That(files2, Has.Count.EqualTo(1));
-            Assert.That(NonLockingFileReader.ReadAllTextWithoutLocking(files2.First()), Is.EqualTo($"Foo{Environment.NewLine}Bar{Environment.NewLine}"));
+            Assert.That(NonLockingFileReader.ReadAllTextWithoutLocking(files2[0]), Is.EqualTo($"Foo{Environment.NewLine}Bar{Environment.NewLine}"));
         }
     }
 
     [Test]
     public void When_file_is_deleted_underneath_continues_to_write_afterwards()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var logger = new RollingLogger(tempPath.TempDirectory)
         {
             GetDate = () => new DateTimeOffset(2010, 10, 1, 0, 0, 0, TimeSpan.Zero)
@@ -56,7 +56,7 @@ public class RollingLoggerTests
     [Test]
     public void When_file_is_locked_exception_is_swallowed()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var logger = new RollingLogger(tempPath.TempDirectory)
         {
             GetDate = () => new DateTimeOffset(2010, 10, 1, 0, 0, 0, TimeSpan.Zero)
@@ -75,7 +75,7 @@ public class RollingLoggerTests
     [Test]
     public void When_file_is_deleted_underneath_immediately_before_write()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var logger = new RollingLoggerThatDeletesBeforeWrite(tempPath.TempDirectory)
         {
             GetDate = () => new DateTimeOffset(2010, 10, 1, 0, 0, 0, TimeSpan.Zero)
@@ -99,7 +99,7 @@ public class RollingLoggerTests
     [Test]
     public void When_file_already_exists_and_is_too_large_a_new_sequence_file_is_written()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var utcOffset = TimeZoneInfo.Local.GetUtcOffset(DateTime.UtcNow);
 
         var dateTime = new DateTimeOffset(2010, 10, 1, 0, 0, 0, utcOffset);
@@ -121,7 +121,7 @@ public class RollingLoggerTests
         using (Assert.EnterMultipleScope())
         {
             Assert.That(Path.GetFileName(first), Is.EqualTo("nsb_log_2010-10-01_0.txt"));
-            Assert.That(NonLockingFileReader.ReadAllTextWithoutLocking(files.First()), Is.EqualTo($"Some long text{Environment.NewLine}"));
+            Assert.That(NonLockingFileReader.ReadAllTextWithoutLocking(first), Is.EqualTo($"Some long text{Environment.NewLine}"));
         }
 
         var second = files[1];
@@ -135,7 +135,7 @@ public class RollingLoggerTests
     [Test]
     public void When_file_already_exists_with_wrong_date_a_file_is_written()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var logger1 = new RollingLogger(tempPath.TempDirectory)
         {
             GetDate = () => new DateTimeOffset(2010, 10, 1, 0, 0, 0, TimeSpan.Zero)
@@ -154,7 +154,7 @@ public class RollingLoggerTests
         using (Assert.EnterMultipleScope())
         {
             Assert.That(Path.GetFileName(first), Is.EqualTo("nsb_log_2010-10-01_0.txt"));
-            Assert.That(NonLockingFileReader.ReadAllTextWithoutLocking(files.First()), Is.EqualTo($"Foo{Environment.NewLine}"));
+            Assert.That(NonLockingFileReader.ReadAllTextWithoutLocking(first), Is.EqualTo($"Foo{Environment.NewLine}"));
         }
 
         var second = files[1];
@@ -168,7 +168,7 @@ public class RollingLoggerTests
     [Test]
     public void When_line_is_write_line_appears_in_file()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var logger = new RollingLogger(tempPath.TempDirectory);
         logger.WriteLine("Foo");
         var singleFile = tempPath.GetSingle();
@@ -178,7 +178,7 @@ public class RollingLoggerTests
     [Test]
     public void When_multiple_lines_are_written_lines_appears_in_file()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var logger = new RollingLogger(tempPath.TempDirectory);
         logger.WriteLine("Foo");
         logger.WriteLine("Bar");
@@ -189,7 +189,7 @@ public class RollingLoggerTests
     [Test]
     public void When_max_file_size_is_exceeded_sequence_number_is_added()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var utcOffset = TimeZoneInfo.Local.GetUtcOffset(DateTime.UtcNow);
         var logger = new RollingLogger(tempPath.TempDirectory, maxFileSize: 10)
         {
@@ -218,7 +218,7 @@ public class RollingLoggerTests
     [Test]
     public void When_many_sequence_files_are_written_the_max_is_not_exceeded()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var logger = new RollingLogger(tempPath.TempDirectory, maxFileSize: 10)
         {
             GetDate = () => new DateTimeOffset(2010, 10, 1, 0, 0, 0, TimeSpan.Zero)
@@ -233,7 +233,7 @@ public class RollingLoggerTests
     [Test]
     public void When_new_write_causes_overlap_of_file_size_line_is_written_to_current_file()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var logger = new RollingLogger(tempPath.TempDirectory, maxFileSize: 10);
         logger.WriteLine("Foo");
         logger.WriteLine("Some long text");
@@ -244,7 +244,7 @@ public class RollingLoggerTests
     [Test]
     public void When_date_changes_new_file_is_written()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var logger = new RollingLogger(tempPath.TempDirectory)
         {
             GetDate = () => new DateTimeOffset(2010, 10, 1, 0, 0, 0, TimeSpan.Zero)
@@ -273,7 +273,7 @@ public class RollingLoggerTests
     [Test]
     public void When_getting_todays_log_file_sequence_number_is_used_in_sorting()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var today = new DateTimeOffset(2010, 10, 2, 0, 0, 0, TimeSpan.Zero);
         var logFiles = new List<RollingLogger.LogFile>
         {
@@ -287,7 +287,7 @@ public class RollingLoggerTests
     [Test]
     public void When_getting_todays_log_file_only_today_is_respected()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var today = new DateTimeOffset(2010, 10, 2, 0, 0, 0, TimeSpan.Zero);
         var yesterday = new DateTimeOffset(2010, 10, 1, 0, 0, 0, TimeSpan.Zero);
         var tomorrow = new DateTimeOffset(2010, 10, 3, 0, 0, 0, TimeSpan.Zero);
@@ -306,7 +306,7 @@ public class RollingLoggerTests
     [Test]
     public void When_many_files_written_over_size_old_files_are_deleted()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var utcOffset = TimeZoneInfo.Local.GetUtcOffset(DateTime.UtcNow);
         var logger = new RollingLogger(tempPath.TempDirectory, numberOfArchiveFilesToKeep: 2, maxFileSize: 5)
         {
@@ -345,7 +345,7 @@ public class RollingLoggerTests
     [Test]
     public void When_many_files_written_over_dates_old_files_are_deleted()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var logger = new RollingLogger(tempPath.TempDirectory, numberOfArchiveFilesToKeep: 2)
         {
             GetDate = () => new DateTimeOffset(2010, 10, 1, 0, 0, 0, TimeSpan.Zero)
@@ -387,7 +387,7 @@ public class RollingLoggerTests
     [Test]
     public void When_line_is_write_file_has_correct_name()
     {
-        using var tempPath = new TempPath();
+        using var tempPath = new TempPath("RollingLoggerTests");
         var logger = new RollingLogger(tempPath.TempDirectory)
         {
             GetDate = () => new DateTimeOffset(2010, 10, 1, 0, 0, 0, TimeSpan.Zero)
@@ -395,31 +395,5 @@ public class RollingLoggerTests
         logger.WriteLine("Foo");
         var singleFile = tempPath.GetSingle();
         Assert.That(Path.GetFileName(singleFile), Is.EqualTo("nsb_log_2010-10-01_0.txt"));
-    }
-    class TempPath : IDisposable
-    {
-        public TempPath()
-        {
-            TempDirectory = Path.Combine(Path.GetTempPath(), "nsbLogging", Guid.NewGuid().ToString());
-            Directory.CreateDirectory(TempDirectory);
-        }
-
-        public readonly string TempDirectory;
-
-        public List<string> GetFiles() => [.. Directory.GetFiles(TempDirectory).OrderBy(x => x)];
-
-        public string GetSingle() => GetFiles().Single();
-
-        public void Dispose() => Directory.Delete(TempDirectory, true);
-    }
-
-    static class NonLockingFileReader
-    {
-        internal static string ReadAllTextWithoutLocking(string path)
-        {
-            using var fileStream = new FileStream(path, FileMode.Open, FileAccess.Read, FileShare.ReadWrite);
-            using var textReader = new StreamReader(fileStream);
-            return textReader.ReadToEnd();
-        }
     }
 }
